# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Photography Booking Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /photographers:
    get:
      operationId: getPhotographers
      tags: [photographers]
      summary: List all photographers
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of photographers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photographer"
    post:
      operationId: createPhotographer
      tags: [photographers]
      summary: Create a new photographer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhotographerBody"
      responses:
        "201":
          description: Created photographer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photographer"
  /photographers/{id}:
    get:
      operationId: getPhotographer
      tags: [photographers]
      summary: Get photographer details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photographer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotographerDetail"
    put:
      operationId: updatePhotographer
      tags: [photographers]
      summary: Update photographer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhotographerBody"
      responses:
        "200":
          description: Updated photographer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photographer"
  /photographers/available:
    get:
      operationId: getAvailablePhotographers
      tags: [photographers]
      summary: Get photographers available for a specific date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability grouped list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotographerAvailability"
...
components:
  schemas:
...
    Photographer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialization:
          type: string
        experience:
          type: integer
        status:
          type: string
          enum: [available, busy, unavailable]
        avatarUrl:
          type: ["string", "null"]
        workingDays:
          type: array
          items:
            type: string
        city:
          type: string
        currentBookingCount:
          type: integer
      required:
        - id
        - name
        - email
        - phone
        - specialization
        - experience
        - status
        - workingDays
        - city
        - currentBookingCount
    PhotographerDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialization:
          type: string
        experience:
          type: integer
        status:
          type: string
          enum: [available, busy, unavailable]
        avatarUrl:
          type: ["string", "null"]
        workingDays:
          type: array
          items:
            type: string
        city:
          type: string
        currentBookingCount:
          type: integer
        skills:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        completedShoots:
          type: array
          items:
            $ref: "#/components/schemas/CompletedShoot"
        upcomingBookings:
          type: array
          items:
            $ref: "#/components/schemas/UpcomingBooking"
      required:
        - id
        - name
        - email
        - phone
        - specialization
        - experience
        - status
        - workingDays
        - city
        - currentBookingCount
        - skills
        - completedShoots
        - upcomingBookings
    CompletedShoot:
      type: object
      properties:
        id:
          type: integer
        eventType:
          type: string
        date:
          type: string
        clientName:
          type: string
        location:
          type: string
      required:
        - id
        - eventType
        - date
        - clientName
        - location
    UpcomingBooking:
      type: object
      properties:
        id:
          type: integer
        eventType:
          type: string
        date:
          type: string
        clientName:
          type: string
        location:
          type: string
        status:
          type: string
      required:
        - id
        - eventType
        - date
        - clientName
        - location
        - status
    PhotographerAvailability:
      type: object
      properties:
        available:
          type: array
          items:
            $ref: "#/components/schemas/Photographer"
        unavailable:
          type: array
          items:
            $ref: "#/components/schemas/UnavailablePhotographer"
      required:
        - available
        - unavailable
    UnavailablePhotographer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialization:
          type: string
        avatarUrl:
          type: ["string", "null"]
        reason:
          type: string
      required:
        - id
        - name
        - specialization
        - reason
    CreatePhotographerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialization:
          type: string
        experience:
          type: integer
        workingDays:
          type: array
          items:
            type: string
        city:
          type: string
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
      required:
        - name
        - email
        - phone
        - specialization
        - experience
        - workingDays
        - city
    UpdatePhotographerBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        experience:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        workingDays:
          type: ["array", "null"]
          items:
            type: string
        city:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        skills:
          type: ["array", "null"]
          items:
            type: string
...