# lib/api-zod/src/generated/types/completedShoot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Booking Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface CompletedShoot {
  id: number;
  eventType: string;
  date: string;
  clientName: string;
  location: string;
}